Create Script.js
A 3 function row column checker to compare for repeating elements in arrays. Main function takes the Sudoku array, gets the length and calculates the row, sends it as an argument to the column function and finally the check function. Program returns true or false which triggers some text to the console.